// src/app/features/guest/models/guest-api.model.ts
export default class GuestApiModel {
  title?: string;

  firstName: string;

  lastName: string;

  email: string;

  primaryPhoneNumber?: string;

  secondaryPhoneNumber?: string;

  companyName: string;

  typeOfVisit: string;

  location: string;

  validFromDate: string;

  validFromTime: string;

  validUntilDate: string;

  validUntilTime: string;

  issuedBy?: string;
}